fix millisecond calculation

+++ firmware/src/sys.c
@@ -77,5 +77,5 @@ int64_t sys_get_timestamp() {
   struct timeval tv;
   gettimeofday(&tv, NULL);
 
-  return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+  return (int64_t)(tv.tv_sec) * 1000 + (int64_t)(tv.tv_usec) / 1000;
 }
